chore: some refactor and docfix
