+++ pkg/sources/transformer/grpc_transformer.go
@@ -19,18 +19,18 @@ package transformer
 import (
 	"context"
 	"fmt"
-	"github.com/numaproj/numaflow/pkg/sdkclient/udf/client"
 	"time"
+
+	"google.golang.org/protobuf/types/known/emptypb"
+	"google.golang.org/protobuf/types/known/timestamppb"
 
 	functionpb "github.com/numaproj/numaflow-go/pkg/apis/proto/function/v1"
 	"github.com/numaproj/numaflow/pkg/forward/applier"
 	"github.com/numaproj/numaflow/pkg/isb"
+	udferr "github.com/numaproj/numaflow/pkg/sdkclient/error"
+	"github.com/numaproj/numaflow/pkg/sdkclient/udf/client"
 	"github.com/numaproj/numaflow/pkg/udf/function"
-	"github.com/numaproj/numaflow/pkg/udferr"
 	"k8s.io/apimachinery/pkg/util/wait"
-
-	"google.golang.org/protobuf/types/known/emptypb"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 // gRPCBasedTransformer applies user defined transformer over gRPC (over Unix Domain Socket) client/server where server is the transformer.
